requirements.txt
audioread==2.1.8
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
future==0.18.2
gensim==3.8.3
idna==2.10
joblib==0.16.0
librosa==0.6.3
llvmlite==0.36.0
numba==0.53.1
numpy==1.19.2
Pillow==8.3.1

requests==2.24.0
resampy==0.2.2
scikit-learn==0.23.2
six==1.15.0
sklearn==0.0
smart-open==2.2.0
threadpoolctl==2.1.0
tqdm==4.49.0
urllib3==1.25.10
